requirements.txt

requests
python-dotenv
supabase
python-binance
pandas